Fix type conversion
Multiplier is a signed value (for no apparent reason)
As a result it has to be cast to unsigned. 
C# does not implicitly cast signed to unsigned values.

+++ src/System.Management.Automation/engine/parser/tokenizer.cs
@@ -3261,7 +3261,7 @@ namespace System.Management.Automation.Language
                     {
                         if (UInt32.TryParse(strNum, style, NumberFormatInfo.InvariantInfo, out uint u))
                         {
-                            ulong testresult = u * multiplier;
+                            ulong testresult = u * (ulong)multiplier;
                             if (testresult > UInt32.MaxValue)
                             {
                                 result = testresult;
